feat: Avoid boxing primitives in Go in sequence concatenations as well (#6361)
### What was changed?
Continuing from #6353, this change ensures the same optimization to
avoid boxing bytes, Chars and CodePoints in sequences is also used in
the concatenation optimization implemented in `dafnyRuntime.dfy`. In
particular, it ensures the new `Array`s created to hold the result are
also optimized for those types. This required modifying the
`Sequence<T>` implementation to expose a `PrototypeArray()` to indicate
the array type at runtime. The overall solution mirrors the solution in
`DafnyRuntimeJava` quite closely.

### How has this been tested?
Existing tests. Performance improvement was measured using
https://github.com/aws/aws-encryption-sdk/tree/mainline/esdk-performance-testing/benchmarks/go.
On the `--quick` scenario (a payload of 50 MB), the peak memory usage
before was 5822mb, and after the optimization was 595mb. The average
latency also improved from 3652 ms to 2047 ms.