fix: Go Array performance (#6371)
### What was changed?

There are two issues causing excessive memory usage when using Seq in
Dafny Go:
1. Dafny Go compiler uses an empty sequence whenever declaring a new
variable. It doesn't use the specific type for sequence, but rather an
`arrayStruct` which uses `interface{}` as the underlying type. While the
[Boxing Fix](https://github.com/dafny-lang/dafny/pull/6361) addressed
the issue when converting from native Go types to dafny Go types (and
vice versa), it didn't fix the use of arrayStruct /
[EmptySeq](https://github.com/dafny-lang/DafnyRuntimeGo/blob/38e14aaf73eff74913a994ee80fd49edf5161c17/v4/dafny/dafny.go#L517)
in the compiler.

2. This issue gets compounded when using ConcatSequence for storing byte
sequences. The
[ToByteArray](https://github.com/dafny-lang/DafnyRuntimeGo/blob/38e14aaf73eff74913a994ee80fd49edf5161c17/v4/dafny/dafny.go#L710-L731)
method for concat seq delegates to ToArray which then uses a Vector to
flatten everything into arrayStruct, causing the same issue which was
fixed by the Boxing Fix. But since the ConcatSequence also has to use
ArraySequence as the final underlying type, delegating it to the
ToByteArray recursively ends up calling the optimized ToByteArrray for
ArraySequence, improving the performance from using 800MB for 50MB
sequence to ~100MB.

This PR fixes issue 2 since fixing 2 improves memory numbers AND 3x
boost in cpu efficiency / latency numbers.

NOTE: The numbers can be further optimized by optimizing the compiler to
use specific sequence instead of arrayStruct (Issue 1) but that might be
more involved changed. A PoC compiler fix shows marginal better memory
efficiency than the fix in this PR (which makes sense since recursive
call is expensive memory wise) but worse latency / CPU numbers. A full
fix might bring the CPU numbers in line as well but it's out of scope
for this PR.


---------

Co-authored-by: Shubham Chaturvedi <scchatur@amazon.com>
Co-authored-by: Robin Salkeld <salkeldr@amazon.com>