fix: Statistics library build (#6402)
### What was changed?
In #6374 I missed the fact that the new library wasn't getting picked up
by the TargetSpecific build, because the file names didn't include the
`-cs` tags.

On fixing this a couple of minor verification issues showed up, so also
fixed those. Also found a way to avoid the include statements by adding
`Helpers.dfy` to the TargetSpecific project.

### How has this been tested?
Existing tests (now that they are actually running :)